ml_service/requirements.txt
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
idna==3.11
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
numba==0.64.0
numpy==2.4.3
nvidia-nccl-cu12==2.29.7
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
protobuf==7.34.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
shap==0.51.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.6
xgboost==3.2.0
